Prepare to modify database.

+++ MVC5-Seneca/ViewModels/HfedScheduleViewModel.cs
@@ -48,7 +48,7 @@ namespace MVC5_Seneca.ViewModels
         public string HfedClientIds { get; set; }  // IDs fFor multi-select Clients & DropDownList 
 
         [DisplayName("Volunteer Hours")]
-        public int? VolunteerHours { get; set; }
+        public float? VolunteerHours { get; set; }
 
         public int Location_Id { get; set; }
         public int PointPerson_Id { get; set; }
